linux.requirements.txt
absl-py==1.4.0
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
cryptography==41.0.3
Deprecated==1.2.14
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.101.1
flake8==6.1.0
flatbuffers==23.5.26
future==0.18.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.56.2
greenlet==2.0.2
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
jaeger-client==4.8.0
joblib==1.3.2
jose==1.0.0
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.7.0
motor==3.2.0
nose2==0.13.0
oauthlib==3.2.2
opentelemetry-api==1.20.0
opentelemetry-exporter-jaeger==1.20.0
opentelemetry-exporter-jaeger-proto-grpc==1.20.0
opentelemetry-exporter-jaeger-thrift==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
opentracing==2.4.0
opentracing-instrumentation==3.3.1
opt-einsum==3.3.0
packaging==23.1
protobuf==3.20.3
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pyflakes==3.1.0
PyJWT==2.8.0
pymongo==4.4.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tensorflow==2.12.0
termcolor==2.3.0
threadloop==1.0.2
threadpoolctl==3.2.0
thrift==0.16.0
typing-extensions>=4
tornado==5.1.1
urllib3==1.26.16
uvicorn==0.23.2
pillow==10.1.0
keras_preprocessing==1.1.2